adl: documented issue with EOF not correct behaviour
This is possibly a CP/M implementation difference.

If BDOS 33 (READ RANDOM) return 01 (read an unwritten file), the current
implementation will interpret this as EOF - even if this happens in the middle of the
file.

Had to modify the FILETEST.BBC code and comment line 550.

This issue also occurs on the upstream original BBCBASIC.COM for Z80,
implying, a CP/M variation not accounted for.